Move esRangeValue
